Telegram_bot_project/requirements.txt
python-telegram-bot==20.1
peewee==3.15.4
requests==2.28.1
python-dotenv==0.21.0
